requirements.txt
openai
shot-scraper
python-dotenv
markdown
weasyprint